Update README.md
In this project, a secret message is hidden in an image and later extracted using the
LSB Image Steganography Technique. The 2 images are cyber3.jpeg and dog.jpeg
and the 2 text files are pattern.txt and test.txt respectively. The steganography
folder has the input file and the output is saved in the folder called scan.